order

+++ app/Http/Controllers/Admin/PostController.php
@@ -54,7 +54,7 @@ class PostController extends Controller
 
     public function edit(Post $post)
     {
-        // $this->authorize('author',$post);
+        $this->authorize('author',$post);
         $categories = Category::pluck('name', 'id');
         $tags = Tag::all();
         return view('admin.posts.edit', compact('post', 'categories', 'tags'));
@@ -95,7 +95,7 @@ class PostController extends Controller
     
     
     public function destroy(Post $post){
-        // $this->authorize('author',$post);
+        $this->authorize('author',$post);
         $post->delete();
         return redirect()->route('admin.posts.index')->with('success','El post ha sido eliminado exitosamente');
     }
